ENH/BUG: add Artifact/Visualization.peek for quick inspection of archives
Solves bug in Artifact/Visualization.load where previously the archive filepath was extracted on demand, resulting in buggy behavior if the filepath was deleted or modified after loading but prior to extraction. This problem is not easily solvable in a cross-platform compatible way so a new `.peek()` classmethod is provided to quickly retrieve an archive's metadata (e.g. UUID, type, and provenance) without extracting. `.load()` now extracts immediately.

Fixes #24.