ENH: Artifacts are coerced by transformers
Pair-programmed with @thermokarst

Large change which has removed readers/writers for data-layouts. Instead
data-layouts are now called DirectoryFormats which have no version or name,
instead the classname is used. DirectoryFormats are composed of
FileFormats, which basically match the existing class.

Transformers are registered on the plugin and indicate their ability to
convert one type into another via function annotations.

DirectoryFormats and FileFormats can also be requested as a view and
provide the path as `obj.path` and as `str(obj)`. DirectoryFormats have
additional API which should be considered a WIP.