REF: Method/Visualizer -> Action (#118)
Refactored qiime.sdk.Method and qiime.sdk.Visualizer, which previously contained a ton of duplicated code. The classes were also monolithic.

The new design removes Method and Visualizer classes in favor of a general Action class, which wraps a Method/VisualizerCallable. The Callable wraps an appropriate Signature type and is responsible for synchronous/asynchronous execution (including pickling). `from_function`/`from_markdown` now defer to the relevant components for processing these sources. Plugin method/visualizer registration acts as a factory that builds an appropriate Action and registers it with the framework.

The execution of an `Action` (i.e. the "artifact API") returns a `Results` object containing named results (currently only when there is more than one output; this behavior may change in the future).