ENH: Archive version 0.3.0: use archive UUID as zipfile root directory
Archive version 0.3.0 is not backward-compatible with older versions. The new format uses the archive's UUID as the root directory in the zipfile. Thus, extracting an archive will create a single directory with the archive's UUID as its name.

Previous archive formats determined the root directory name from the filepath the archive was being loaded from or saved to. This was fragile for two reasons: 1) if the file is renamed after saving, it became unloadable (!!!), and 2) if the file was extracted (e.g. using `unzip`), it could potentially mangle an existing directory that happened to have the same name. Now that a UUID is used for the extracted directory, collisions like this should never happen in practice.

Many thanks to @thermokarst for having the idea to use the archive's UUID as its root directory!

Other changes/notes:

- When loading an archive, dotfiles/dirs are ignored so that accidental addition of these files (e.g. user unzips, creates a .DS_Store, and rezips) won't make the archive unloadable.

- Archiver now manages UUID parsing, formatting, validation, and generation. UUID concerns were previously shared between Archiver, Artifact, and Visualization. UUIDs are now assured to be valid version 4 UUIDs, where other types of UUIDs were inadverently allowed before.  These changes to Archiver actually simplified code by removing constructor parameter `archive_filepath`, which was mutually exclusive with `data_initializer`. Now, Archiver requires a `data_initializer`.  If a UUID is provided to the constructor, it will be used (e.g. when loading an archive), otherwise a random version 4 UUID is generated by Archiver. These changes required modifications to Archiver's constructor signature (it's a private class though so shouldn't impact users).

Fixes #140.