REF: Refactored Archiver logic (#165)
Archiver has become its own subpackage `archive`. There are now 3 distinct
components: `Archiver`, `_Archive` impl, and `ArchiveFormat`.

The `Archiver` is responsible for coordinating, it maintains a list of all
current archive formats and archive implementations as well as what the current
version of both of those are.

An `_Archive` implementation is responsible for the high level structure of an
archive (the uuid root dir and VERSION file). Its job is to recognize if a
given file is valid archive and if so mount/extract it. Additionally it
provides a `setup` which initializes an existing directory.

An `ArchiveFormat` specifies the rest of the layout of an archive, this
includes what the data directory is called, metadata, etc...

tldr; Archiver gets ZipArchive to convert an archive to the filesystem and then
gets ArchiveFormat to read the details.

Additionally it is now possible to `peek` at an archive even if you do not have
the correct plugins installed to `load` an archive.

Similarly you can `extract` an archive even if you do not have the latest
ArchiveFormat.