ENH: Adds basic provenance parsing support
parse_actions is a private utility helper that accepts v0- and v1- format
Artifacts, and returns a NetworkX DiGraph where the nodes represent artifacts
and the edges represent actions. The nodes and edges each carry the necessary
attributes (as extracted from the provenance) to recreate the execution
pipeline necessary to derive the  artifact being processed.
This utility is intended to be a stop-gap measure for now, and we anticipate
replacing it with a more full-featured provenance object.