ENH: remove `qiime2.plugin.plugin_init` in favor of direct Cookiecutter usage (#211)
The API was only a thin wrapper around Cookiecutter, which already has a Python API and CLI. It's easier to have plugin devs use Cookiecutter directly when templating a plugin, and lets us remove Cookuecutter as a framework dependency. There are instructions on the QIIME 2 docs site for using Cookiecutter.