ENH: qiime2.sdk.PluginManager.importable_types; sandboxed tests (#216)
Adds `qiime2.sdk.PluginManager.importable_types` read-only property, which is the set of concrete semantic types that are importable (i.e.  concrete type expressions that are registered with a directory format).

Also made testing sandboxed, such that when the QIIMETEST environment variable is set, only the framework's builtin `dummy-plugin` is discovered. If QIIMETEST is not set, all available plugins are discovered, with the exception of `dummy-plugin`.

This sandboxing solves a couple of issues:

- Unit tests are more robust and easier to write, because when tests are run, only the framework's `dummy-plugin` is discovered.  Previously tests had to be written so they would pass even if other plugins were discovered.

- q2-dummy-types and the framework's `dummy-plugin` are incompatible because they define types with name conflicts. Previously it was not possible to run tests with QIIMETEST and have q2-dummy-types installed at the same time. Now, only one plugin is loaded depending on whether tests are being run or not.