BUG: fix SingleFileDirectoryFormat transformation (#263)
When importing it is sometimes the case that the user is importing a directory
and both the supplied view_type and the registered view_type are
SingleFileDirectoryFormats with only a single transformer between their
respective file-formats. This should be fine, but the implicit transformers of:

XDirFmt -> X => Y -> YDirFmt

were not working (where -> is an implicit transformer and => is a registered
one). Previously they would raise an error.