ENH: support optional artifacts (#289)
Methods and visualizers now support optional artifacts. If an artifact is optional the only supported default value is `None`.

With this change, signature inputs and parameters are now allowed to be declared in any order (previously inputs were required to be followed by parameters, and they were not allowed to be mixed).

Removed `Signature.defaults` property in favor of using the `ParameterSpec` objects available within `Signature`.

Fixes #235.