Update primitive.py for range to support
Currently, the repr for a range [1:] with no upper value is the same as the repr for a range [:1] with no lower value.

In [1]: from qiime2.core.type.primitive import Range
In [2]: repr(Range(1, None)) == repr(Range(None, 1))
Out[2]: True

Proposed change forces the repr to include both start and end positions, even if they are the same as the default. This will allow the repr to differentiate between the two ranges in the case above (as when used in parse_type to generate a type from a repr string).