BUG: Fix repr of Range to support start and end correctly (#339)
This commit forces the repr to include both start and end positions, even if they are the same as the default. This will allow the repr to differentiate between the two ranges in the case above (as when used in parse_type to generate a type from a repr string).