BUG: remove all identity predicates
When Choices was changed to disallow an empty set, it introduced a bug
where `Str >= Str % Choices({...})` would raise an error instead of
reporting `True`. This expression never really happens in practice, so
it doesn't impact much of anything. But now there are more generic
protections preventing a plugin developer from registering an empty
predicate, which is probably not what they intended.