ENH: `Metadata.merge`: disallow merging nothing
Previous behavior of `Metadata.merge` allowed zero `Metadata` objects as input, which would result in a copy of the `Metadata` object being called upon (i.e. `self`).

New behavior is to raise a `ValueError` if no `Metadata` objects are provided to merge into `self` (there is nothing to merge in that case).

@thermokarst found a bug in q2cli related to this behavior: a merge was always happening even if only a single `Metadata` object was provided as input, making the `id_header` normalize to `'id'`, which is probably not what users would expect to happen.