NEW: adds checksums to archive structure
This commit adds a 'checksums.md5' file to the root of an archive.
The file is formatted with the md5sum format:

<hex-hash><space><space or asterisk><escaped filename><newline>

Filename escaping handles backslash and newline. In the event those are
present, the backslash is doubled and the newline becomes `\n`.
Additionally if escaping is required, then an additional backslash
precedes the <hex-hash> to indicate the filename is escaped.
See:
https://www.gnu.org/software/coreutils/manual/html_node/md5sum-invocation.html
for more details.