adapt test_metadata_column to accept int64
On 64-bits CPUs, the pandas.dtype returned for integers is a plain
'int' 64 bits wide.  On 32-bits machines however, the type is returned
as 'int64' to differenciate from plain 'int', which is 32 bits wide in
such context.

When running the test suite on 32-bits CPU architectures, the error
output includes several instances of the following:

	======================================================================
	ERROR: test_artifacts_mismatch (qiime2.metadata.tests.test_metadata_column.TestEqualityOperators)
	----------------------------------------------------------------------
	Traceback (most recent call last):
	  File "/tmp/autopkgtest-lxc.shmdvuw3/downtmp/autopkgtest_tmp/qiime2/metadata/tests/test_metadata_column.py", line 344, in test_artifacts_mismatch
	    mdc1 = DummyMetadataColumn(series)
	  File "/tmp/autopkgtest-lxc.shmdvuw3/downtmp/autopkgtest_tmp/qiime2/metadata/metadata.py", line 875, in __init__
	    raise TypeError(
	TypeError: DummyMetadataColumn 'col1' does not support a pandas.Series object with dtype int64

This commit ought to enable the 'int64' type for the class of the test
suite DummyMetadataColumn.